Functions

Vertices

MeshCore.Vertices — Method

Vertices(xyz::Array{T, 2}) where {T}

Construct from a matrix (2D array), one vertex per row.

MeshCore.nvertices — Function

nvertices(v::Vertices{N, T}) where {N, T}

Retrieve number of vertices.

nvertices(sd::AbstractShapeDesc{MD, NV, NF, N1OV, FD}) where {MD, NV, NF, N1OV, FD}

How many vertices does the shape connect?

nvertices(shapes::ShapeCollection)

Retrieve the number of vertices per shape.

MeshCore.nspacedims — Function

nspacedims(v::Vertices{N, T}) where {N, T}

Retrieve the number of space dimensions.

MeshCore.coordinatetype — Function

coordinatetype(v::Vertices{N, T}) where {N, T}

Retrieve type of individual coordinates.

MeshCore.coordinates — Function

coordinates(vertices::Vertices{N, T}, i::I) where {N, T, I<:Int}

Retrieve coordinates of a single vertex.

coordinates(vertices::Vertices{N, T}, I::SVector) where {N, T}

Access coordinates of selected vertices.

Shapes

MeshCore.ShapeCollection — Method

ShapeCollection(shapedesc::S, C::Array{T, 2}) where {S <: AbstractShapeDesc, T}

Convenience constructor from a matrix. One shape per row.

MeshCore.shapedesc — Function

shapedesc(shapes::ShapeCollection)

Retrieve the shape descriptor.

MeshCore.connectivity — Function

connectivity(shapes::ShapeCollection, i::I) where {I}

Retrieve connectivity of one shape from the collection.

connectivity(shapes::ShapeCollection, I::SVector)

Retrieve connectivity of multiple shapes from the collection.

Static arrays are used to help the compiler avoid memory allocation.

MeshCore.nshapes — Function

nshapes(shapes::ShapeCollection)

Number of shapes in the collection.

MeshCore.manifdim — Function

manifdim(sd::AbstractShapeDesc{MD, NV, NF, N1OV, FD}) where {MD, NV, NF, N1OV, FD}

What is the manifold dimension of this shape?

manifdim(shapes::ShapeCollection)

Retrieve the manifold dimension of the collection.

Missing docstring.

Missing docstring for nvertices. Check Documenter's build log for details.

MeshCore.n1storderv — Function

nfacets(sd::AbstractShapeDesc{MD, NV, NF, N1OV, FD}) where {MD, NV, NF, N1OV, FD}

How many 1st-order vertices is there per facet?

For instance, for hexahedra each facet has 4 1st-order vertices.

MeshCore.facetdesc — Function

facetdesc(shapes::ShapeCollection)

Retrieve the shape type of the boundary facet.

MeshCore.nfacets — Function

nfacets(sd::AbstractShapeDesc{MD, NV, NF, N1OV, FD}) where {MD, NV, NF, N1OV, FD}

How many facets bound the shape?

nfacets(shapes::ShapeCollection)

Retrieve the number of boundary facets per shape.

MeshCore.facets — Function

facets(shapes::ShapeCollection)

Retrieve the connectivity of the facets.

MeshCore.facetconnectivity — Function

connectivity(shapes::ShapeCollection, i::I) where {I}

Retrieve connectivity of one shape from the collection.

Topology operations

MeshCore.skeleton — Function

skeleton(shapes::ShapeCollection; options...)

Compute the skeleton of the shape collection.

This computes a new shape collection from an existing shape collection. It consists of facets (shapes of manifold dimension one less than the manifold dimension of the shapes themselves).

Options

boundaryonly: include in the skeleton only shapes on the boundary of the input collection, true or false (default).

MeshCore.boundary — Function

boundary(shapes::ShapeCollection)

Compute the shape collection for the boundary of the collection on input.

This is a convenience version of the skeleton function.

MeshCore.increl_vertextoshape — Function

increl_vertextoshape(shapes::ShapeCollection)

Compute the incidence relation 0 -> d for d-dimensional shapes.

This only makes sense for d > 0. For d=1 we get for each vertex the list of edges connected to the vertex, and analogously faces and cells for d=2 and d=3.

Missing docstring.

Missing docstring for (ir::IncRelVertexToShape)(j::Int64). Check Documenter's build log for details.

MeshCore.increl_boundedby — Function

increl_boundedby(shapes::ShapeCollection, facetshapes::ShapeCollection)

Compute the incidence relation d -> d-1 for d-dimensional shapes.

In other words, this is the incidence between shapes and the shapes that bound these shapes (facets). For tetrahedra as the shapes, the incidence relation lists the numbers of the faces that bound each individual tetrahedron.

!!! Note The numbers of the facets are signed: positive when the facet bounds the shape in the sense in which it is defined by the shape as oriented with an outer normal; negative otherwise. The sense is defined by the numbering of the 1st-order vertices of the facet shape.

Missing docstring.

Missing docstring for (ir::IncRelBoundedBy)(j::Int64). Check Documenter's build log for details.

Index

MeshCore.AbstractShapeDesc

MeshCore.H8ShapeDesc

MeshCore.IncRelBoundedBy

MeshCore.IncRelVertexToShape

MeshCore.L2ShapeDesc

MeshCore.NoSuchShapeDesc

MeshCore.P1ShapeDesc

MeshCore.Q4ShapeDesc

MeshCore.ShapeCollection

MeshCore.ShapeCollection

MeshCore.T3ShapeDesc

MeshCore.T4ShapeDesc

MeshCore.Vertices

MeshCore.Vertices

MeshCore.boundary

MeshCore.connectivity

MeshCore.coordinates

MeshCore.coordinatetype

MeshCore.facetconnectivity

MeshCore.facetdesc

MeshCore.facets

MeshCore.increl_boundedby

MeshCore.increl_vertextoshape

MeshCore.manifdim

MeshCore.n1storderv

MeshCore.nfacets

MeshCore.nshapes

MeshCore.nspacedims

MeshCore.nvertices

MeshCore.shapedesc

MeshCore.skeleton