Functions

Shapes descriptors

MeshCore.manifdim — Method

manifdim(sd::AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}) where {MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

What is the manifold dimension of this shape?

MeshCore.nvertices — Method

nvertices(sd::AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}) where {MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

How many vertices does the shape connect?

MeshCore.nfacets — Method

nfacets(sd::AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}) where {MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

How many facets bound the shape?

MeshCore.nridges — Method

nridges(sd::AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}) where {MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

How many ridges bound the shape?

MeshCore.n1storderv — Method

n1storderv(sd::AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}) where {MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

How many 1st-order vertices are there per shape?

For instance, for hexahedra each shape has 8 1st-order vertices.

MeshCore.nshifts — Method

nshifts(sd::AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}) where {MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

How many circular shifts should be tried to figure out the orientation (sense)?

MeshCore.nfeatofdim — Function

nfeatofdim(sd::SD, m) where {SD<:AbsShapeDesc}

Compute number of manifold features of given dimension.

For instance, a tetrahedron with four vertices has 1 feature of manifold dimension 3, four features of manifold dimension 2, six features of manifold dimension 1, and four features of manifold dimension 0.

Quadratic or higher order elements may have more vertices than would correspond to the hierarchy of cells, facets, ridges. Internal vertices must be accounted for as well. For instance, L3 has two facets, but three vertices. T6 has three facets, three ridges, but six vertices.

Shapes

MeshCore.ShapeColl — Method

ShapeColl(shapedesc::S, nshapes::Int64) where {S<:AbsShapeDesc}

Set up new shape collection with an empty dictionary of attributes and a default name.

MeshCore.ShapeColl — Method

ShapeColl(shapedesc::S, nshapes::Int64, s::String) where {S<:AbsShapeDesc}

Set up new shape collection with an empty dictionary of attributes.

MeshCore.shapedesc — Function

shapedesc(shapes::ShapeColl)

Retrieve the shape descriptor.

MeshCore.nshapes — Function

nshapes(shapes::ShapeColl)

Number of shapes in the collection.

MeshCore.manifdim — Method

manifdim(shapes::ShapeColl)

Retrieve the manifold dimension of the collection.

MeshCore.nvertices — Method

nvertices(shapes::ShapeColl)

Retrieve the number of vertices per shape.

MeshCore.facetdesc — Function

facetdesc(shapes::ShapeColl)

Retrieve the shape type of the boundary facet.

MeshCore.nfacets — Method

nfacets(shapes::ShapeColl)

Retrieve the number of boundary facets per shape.

MeshCore.facetconnectivity — Function

facetconnectivity(shapes::ShapeColl, j::Int64)

Retrieve connectivity of the j-th facet shape of the i-th shape from the collection.

MeshCore.ridgedesc — Function

ridgedesc(shapes::ShapeColl)

Retrieve the shape type of the boundary ridge.

MeshCore.nridges — Method

nridges(shapes::ShapeColl)

Retrieve the number of boundary ridges per shape.

MeshCore.ridgeconnectivity — Function

ridgeconnectivity(shapes::ShapeColl, j::Int64)

Retrieve connectivity of the j-th ridge shape of the i-th shape from the collection.

MeshCore.attribute — Function

attribute(shapes::ShapeColl, s::Symbol)

Retrieve a named attribute.

Incidence relations

MeshCore.IncRel — Method

IncRel(left::ShapeColl{LEFT}, right::ShapeColl{RIGHT}, v::Vector{T}) where {LEFT<:AbsShapeDesc,RIGHT<:AbsShapeDesc,T}

Convenience constructor with a vector of vectors and a default name.

MeshCore.IncRel — Method

IncRel(left::ShapeColl{LEFT}, right::ShapeColl{RIGHT}, v::Vector{T}, name::String) where {LEFT<:AbsShapeDesc,RIGHT<:AbsShapeDesc,T}

Convenience constructor with a vector of vectors and a name.

MeshCore.IncRel — Method

IncRel(left::ShapeColl{LEFT}, right::ShapeColl{RIGHT}, data::Matrix{MT}) where {LEFT<:AbsShapeDesc,RIGHT<:AbsShapeDesc,MT}

Convenience constructor supplying a matrix instead of a vector of vectors and a default name.

MeshCore.IncRel — Method

IncRel(left::ShapeColl{LEFT}, right::ShapeColl{RIGHT}, data::Matrix{MT}, name::String) where {LEFT<:AbsShapeDesc,RIGHT<:AbsShapeDesc,MT}

Convenience constructor supplying a matrix instead of a vector of vectors and a name.

MeshCore.indextype — Function

indextype(ir::IncRel)

Provide type of the incidence index (entity number). Presumably some type of integer.

MeshCore.nrelations — Function

nrelations(ir::IncRel)

Number of individual relations in the incidence relation.

MeshCore.nentities — Function

nentities(ir::IncRel, j::IT) where {IT}

Number of individual entities in the j-th relation in the incidence relation.

MeshCore.ir_code — Function

code(ir::IncRel{LEFT,RIGHT,T}) where {LEFT<:AbsShapeDesc,RIGHT<:AbsShapeDesc,T}

Formulate the code of the incidence relation.

Relations below the diagonal

MeshCore.ir_skeleton — Function

ir_skeleton(ir::IncRel, name = "skt")

Compute the "skeleton" of an incidence relation.

Compute a new incidence relation from an existing incidence relation to extract the skeleton.

Returns

Incidence relation for the skeleton of the input incidence relation. The left shape collection consists of facets (shapes of manifold dimension one less than the manifold dimension of the shapes themselves). The right shape collection is the same as for the input.

MeshCore.ir_bbyfacets — Function

ir_bbyfacets(ir::IncRel, fir::IncRel, tfir::IncRel, name = "bbf")

Compute the "bounded by facets" incidence relation (d, d-1) for d-dimensional shapes.

In other words, this is the incidence relation between shapes and the shapes that bound these shapes (facets). For tetrahedra as the shapes, the incidence relation lists the numbers of the faces that bound each individual tetrahedron. The resulting left shape is of the same shape description as in the ir.

Arguments

ir = incidence relation (d, 0),

fir = incidence relation (d-1, 0),

tfir = transpose of the above, incidence relation (0, d-1).

Returns

Incidence relation for the bounded-by relationship between shape collections. The left shape collection is the same as for the ir, the right shape collection is for the facets (shapes of manifold dimension one less than the manifold dimension of the shapes themselves).

Note

The numbers of the facets are signed: positive when the facet bounds the shape in the sense in which it is defined by the shape as oriented with an outer normal; negative otherwise. The sense is defined by the numbering of the 1st-order vertices of the facet shape.

ir_bbyfacets(ir::IncRel, fir::IncRel, name = "bbf")

Compute the "bounded by facets" incidence relation (d, d-1) for d-dimensional shapes.

Convenience function where the transpose of the incidence relation on the right is computed on the fly.

See also: ir_bbyfacets(ir::IncRel, fir::IncRel, tfir::IncRel)

MeshCore.ir_bbyridges — Function

ir_bbyridges(ir::IncRel, eir::IncRel, teir::IncRel, name = "bbr")

Compute the "bounded by ridges" incidence relation (d, d-2) for d-dimensional shapes.

In other words, this is the incidence between shapes and the shapes that "bound" the boundaries of these shapes (i. e. ridges). For tetrahedra as the shapes, the incidence relation lists the numbers of the edges that "bound" each individual tetrahedron. The resulting shape is of the same shape description as the shapes on input.

Returns

The incidence relation for the "bounded by" relationship between shape collections. The left shape collection is the same as for the ir, the right shape collection is for the ridges (shapes of manifold dimension two less than the manifold dimension of the shapes themselves).

Note

The numbers of the ridges are signed: positive when the ridge bounds the shape in the sense in which it is defined by the shape as oriented with an outer normal; negative otherwise. The sense is defined by the numbering of the 1st-order vertices of the ridge shape.

ir_bbyridges(ir::IncRel, eir::IncRel, name = "bbr")

Compute the "bounded-by-ridges" incidence relation (d, d-2).

Convenience function where the transpose of the incidence relation on the right is computed on the fly.

See also: ir_bbyridges(ir::IncRel, eir::IncRel, efir::IncRel)

Relations above the diagonal

MeshCore.ir_transpose — Method

ir_transpose(ir::IncRel, name = "trp")

Compute the "transposed" incidence relation (d1, d2), where d2 >= d1.

This only makes sense for d2 >= 0. For d2=1 we get for each vertex the list of edges connected to the vertex, and analogously faces and cells for d=2 and d=3.

Returns

Incidence relation for the transposed incidence relation. The left and right shape collection are swapped in the output relative to the input.

Relations on the diagonal

MeshCore.ir_identity — Function

ir_identity(ir::IncRel, side = :left)

Compute the "identity" incidence relation (d, d).

Either the left (side = :left) or the right (side = :right) shape is mapped to itself.

Utilities

MeshCore.ir_subset — Function

ir_subset(ir::IncRel, list)

Compute the incidence relation representing a subset of the shape collection on the left.

MeshCore.ir_boundary — Function

ir_boundary(ir::IncRel, name = "bdr")

Compute the incidence relation for the boundary of the incidence relation on input.

The skeleton function.

Index of functions

MeshCore.SHAPE_DESC

MeshCore.AbsAttrib

MeshCore.AbsShapeDesc

MeshCore.H8ShapeDesc

MeshCore.IncRel

MeshCore.IncRel

MeshCore.IncRel

MeshCore.IncRel

MeshCore.IncRel

MeshCore.L2ShapeDesc

MeshCore.L3ShapeDesc

MeshCore.NoSuchShapeDesc

MeshCore.P1ShapeDesc

MeshCore.Q4ShapeDesc

MeshCore.Q8ShapeDesc

MeshCore.ShapeColl

MeshCore.ShapeColl

MeshCore.ShapeColl

MeshCore.T3ShapeDesc

MeshCore.T4ShapeDesc

MeshCore.T6ShapeDesc

MeshCore.VecAttrib

MeshCore.attribute

MeshCore.facetconnectivity

MeshCore.facetdesc

MeshCore.indextype

MeshCore.ir_bbyfacets

MeshCore.ir_bbyridges

MeshCore.ir_boundary

MeshCore.ir_code

MeshCore.ir_identity

MeshCore.ir_skeleton

MeshCore.ir_subset

MeshCore.ir_transpose

MeshCore.manifdim

MeshCore.manifdim

MeshCore.n1storderv

MeshCore.nentities

MeshCore.nfacets

MeshCore.nfacets

MeshCore.nfeatofdim

MeshCore.nrelations

MeshCore.nridges

MeshCore.nridges

MeshCore.nshapes

MeshCore.nshifts

MeshCore.nvertices

MeshCore.nvertices

MeshCore.ridgeconnectivity

MeshCore.ridgedesc

MeshCore.shapedesc

Types

Shape descriptors

MeshCore.AbsShapeDesc — Type

AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

Abstract shape descriptor (simplex, or cuboid, of various manifold dimension)

The definitions of concrete types below define a set of shape descriptors, covering the usual element shapes: point, line segment, quadrilateral, hexahedron, and a similar hierarchy for the simplex elements: point, line segment, triangle, tetrahedron.

The concrete types of the shape descriptor will provide access to the following properties of the shape S:

manifdim = manifold dimension of the shape S (i. e. 0, 1, 2, or 3),

nvertices = number of vertices of the shape S,

n1storderv = number of first-order vertices for the shape S, for instance a 6-node triangle has 3 first-order vertices,

nfacets = number of shapes on the boundary of the shape S,

facetdesc = shape descriptor of the shapes on the boundary of the shape S,

facets = definitions of the shapes on the boundary of the shape S in terms of local connectivity

ridgedesc = shape descriptor of the shapes on the boundary of the boundary of the shape S (which we call ridges here),

ridges = definitions of the shapes on the boundary of the boundary of the shape S in terms of local connectivity

The bit-type values are defined with the type parameters:

MD is the manifold dimension,

NV is the number of connected vertices,

NF is the number of boundary facets,

FD is the descriptor of the facet (boundary shape).

NR is the number of boundary ridges,

RD is the descriptor of the ridge (boundary shape).

NFOV is number of first-order vertices,

NSHIFTS is the number of shifts that should be attempted when matching shapes to determine orientation.

MeshCore.NoSuchShapeDesc — Type

NoSuchShapeDesc

Base descriptor: no shape is of this description.

MeshCore.P1ShapeDesc — Type

P1ShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS} <: AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

Shape descriptor of a point shape.

MeshCore.L2ShapeDesc — Type

L2ShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS} <: AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

Shape descriptor of a line segment shape.

MeshCore.L3ShapeDesc — Type

L3ShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS} <: AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

Shape descriptor of a quadratic line segment shape.

MeshCore.Q4ShapeDesc — Type

Q4ShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS} <: AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

Shape descriptor of a quadrilateral shape.

Facets: First the two orthogonal to xi, then the two orthogonal to eta. The coordinate along the edge increases.

Ridges: Counterclockwise.

MeshCore.Q8ShapeDesc — Type

Q8ShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS} <: AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

Shape descriptor of a quadrilateral shape.

Facets: First the two orthogonal to xi, then the two orthogonal to eta. The coordinate along the edge increases.

Ridges: Counterclockwise.

MeshCore.H8ShapeDesc — Type

H8ShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS} <: AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

Shape descriptor of a hexahedral shape.

Facets: First the two faces orthogonal to xi, then the two orthogonal to eta, finally the two orthogonal to zeta. The vertices are given counterclockwise when looking from the outside.

Ridges: First the four edges parallel to xi, then the four parallel to eta, and finally the four parallel to zeta.

MeshCore.T3ShapeDesc — Type

T3ShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS} <: AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

Shape descriptor of a triangular shape.

Facets: We start with the edge opposite to vertex 1, then the edge opposite to vertex 2 and so on. The vertices define counterclockwise orientation.

Ridges: The vertices in counterclockwise order.

MeshCore.T6ShapeDesc — Type

T6ShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS} <: AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

Shape descriptor of a quadratic triangular shape.

Facets: We start with the edge opposite to vertex 1, then the edge opposite to vertex 2 and so on. The vertices define counterclockwise orientation.

Ridges: The vertices in counterclockwise order.

MeshCore.T4ShapeDesc — Type

T4ShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS} <: AbsShapeDesc{MD,NV,NF,FD,NR,RD,NFOV,NSHIFTS}

Shape descriptor of a tetrahedral shape.

Facets: We start with the face opposite to vertex 1, then the face opposite to vertex 2 and so on. The vertices define counterclockwise orientation when viewed from outside of the shape.

Ridges: The vertices of the edges are defined in the order based on the right-hand side rotation rule: rotate about an edge (ridge) as if it was a rotation vector with positive orientation. The vertices on the edge opposite across the tetrahedron are traversed in the order given by the rotation.

MeshCore.SHAPE_DESC — Constant

SHAPE_DESC

Dictionary of all the descriptors.

Shape collections

MeshCore.ShapeColl — Type

ShapeColl{S<:AbsShapeDesc}

This is the type of a homogeneous collection of finite element shapes.

S = shape descriptor: subtype of AbsShapeDesc.

Attributes

MeshCore.AbsAttrib — Type

AbsAttrib{T} <: AbstractArray{T,1}

Abstract type of attribute. It is a subtype of AbstractArray.

MeshCore.VecAttrib — Type

VecAttrib{T} <: AbsAttrib{T}

Attribute stores a quantity of type T, one value per shape. The value is indexed to the serial number of the object.

Example

Attribute to allow access to the locations of the vertices.

using Test
using StaticArrays
using MeshCore: VecAttrib, ShapeColl, P1, attribute
xyz = [0.0 0.0; 633.3333333333334 0.0; 1266.6666666666667 0.0]
N, T = size(xyz, 2), eltype(xyz)
locs =  VecAttrib([SVector{N,T}(xyz[i, :]) for i in 1:size(xyz, 1)])
vertices = ShapeColl(P1, size(xyz, 1))
vertices.attributes["geom"] = locs
a = attribute(vertices, "geom")
@test a[2] == [633.3333333333334, 0.0]

Attribute to label the vertices.

using Test
using StaticArrays
using MeshCore: VecAttrib, ShapeColl, P1, attribute
a = VecAttrib([1 for i in 1:size(xyz, 1)])
vertices = ShapeColl(P1, size(xyz, 1))
vertices.attributes["label"] = a
@test a[3] == 1

The vector attribute supports the methods size, getindex, setindex!.

Incidence relations

MeshCore.IncRel — Type

IncRel{LEFT<:AbsShapeDesc,RIGHT<:AbsShapeDesc,T}

Incidence relation expressing connectivity between an entity of the shape on the left and a list of entities of the shape on the right.

The incidence relation may be referred to by its code: (d1, d2), where d1 is the manifold dimension of the shape collection on the left, and d2 is the manifold dimension of the shape collection on the right.

Main operations referred to with the abbreviations used in the paper:

skt = ir_skeleton

trp = ir_transpose

bbf = ir_bbyfacets

bbr = ir_bbyridges

idt = ir_identity

Access to the incidence relation data is through indexing: For instance

ir[55]

provides access to the entire incidence relation ir of the entity 55. The first component of this incidence relation for the entity 55 is

ir[55][1]

or equivalently

ir[55, 1]