Types

Vertices

MeshCore.Vertices — Type

Vertices{N, T}

Vertices of the mesh.

N = number of coordinates, equal to the number of space dimensions,

T = type of the individual coordinates (concrete subtype of AbstractFloat).

Shapes

Shape descriptors

MeshCore.AbstractShapeDesc — Type

AbstractShapeDesc

Abstract shape descriptor (simplex, or cuboid, of various manifold dimension)

The definitions of concrete types below define a set of shape descriptors, covering the usual element shapes: point, line segment, quadrilateral, hexahedron, and a similar hierarchy for the simplex elements: point, line segment, triangle, tetrahedron.

The concrete types of the shape descriptor will provide access to manifdim = manifold dimension of the shape (0, 1, 2, or 3), nvertices = number of vertices of the shape, nfacets = number of shapes on the boundary of this shape, facetdesc = shape descriptor of the shapes on the boundary of this shape, n1storderv = number of first-order vertices for the shape, for instance a 6-node triangle has 3 first-order vertices, facets = definitions of the shapes on the boundary of this shape in terms of local connectivity

The bit-type values are defined with the type parameters: MD is the manifold dimension, NV is the number of connected vertices, NF is the number of boundary facets, N1OV number of first-order vertices, FD is the descriptor of the facet (boundary shape).

MeshCore.NoSuchShapeDesc — Type

NoSuchShapeDesc

Base descriptor: no shape is of this description.

MeshCore.P1ShapeDesc — Type

P1ShapeDesc{MD, NV, NF, N1OV, FD}

Shape descriptor of a point shape.

MeshCore.L2ShapeDesc — Type

L2ShapeDesc{MD, NV, NF, N1OV, FD}

Shape descriptor of a line segment shape.

MeshCore.Q4ShapeDesc — Type

Q4ShapeDesc{MD, NV, NF, N1OV, FD}

Shape descriptor of a quadrilateral shape.

MeshCore.H8ShapeDesc — Type

H8ShapeDesc{MD, NV, NF, N1OV, FD}

Shape descriptor of a hexahedral shape.

MeshCore.T3ShapeDesc — Type

T3ShapeDesc{MD, NV, NF, N1OV, FD}

Shape descriptor of a triangular shape.

MeshCore.T4ShapeDesc — Type

T4ShapeDesc{MD, NV, NF, N1OV, FD}

Shape descriptor of a tetrahedral shape.

Collection of shapes

MeshCore.ShapeCollection — Type

ShapeCollection{S <: AbstractShapeDesc, N, T}

This is the type of a homogeneous collection of finite element shapes.

S = shape descriptor: subtype of AbstractShapeDesc{MD, NV, NF, FD}

IT = type of the vertex numbers (some concrete type of Integer).

Incidence relations

Missing docstring.

Missing docstring for IncRel. Check Documenter's build log for details.