Functions

Vertices

MeshCore.Vertices — Method

Vertices(xyz::Array{T, 2}) where {T}

Construct from a matrix (2D array), one vertex per row.

MeshCore.nvertices — Function

nvertices(v::Vertices{N, T}) where {N, T}

Retrieve number of vertices.

nvertices(sd::AbstractShapeDesc{MD, NV, NF, FD}) where {MD, NV, NF, FD}

How many vertices does the shape connect?

nvertices(shapes::ShapeCollection)

Retrieve the number of vertices per shape.

MeshCore.nspacedims — Function

nspacedims(v::Vertices{N, T}) where {N, T}

Retrieve the number of space dimensions.

MeshCore.coordinatetype — Function

coordinatetype(v::Vertices{N, T}) where {N, T}

Retrieve type of individual coordinates.

MeshCore.coordinates — Function

coordinates(vertices::Vertices{N, T}, i::I) where {N, T, I<:Int}

Retrieve coordinates of a single vertex.

coordinates(vertices::Vertices{N, T}, I::SVector) where {N, T}

Access coordinates of selected vertices.

Shapes

MeshCore.ShapeCollection — Method

ShapeCollection(shapedesc::S, C::Array{T, 2}) where {S <: AbstractShapeDesc, T}

Convenience constructor from a matrix. One shape per row.

MeshCore.shapedesc — Function

shapedesc(shapes::ShapeCollection)

Retrieve the shape descriptor.

MeshCore.connectivity — Function

connectivity(shapes::ShapeCollection, i::I) where {I}

Retrieve connectivity of one shape from the collection.

connectivity(shapes::ShapeCollection, I::SVector)

Retrieve connectivity of multiple shapes from the collection.

Static arrays are used to help the compiler avoid memory allocation.

MeshCore.nshapes — Function

nshapes(shapes::ShapeCollection)

Number of shapes in the collection.

MeshCore.manifdim — Function

manifdim(sd::AbstractShapeDesc{MD, NV, NF, FD}) where {MD, NV, NF, FD}

What is the manifold dimension of this shape?

manifdim(shapes::ShapeCollection)

Retrieve the manifold dimension of the collection.

Missing docstring.

Missing docstring for nnodes. Check Documenter's build log for details.

MeshCore.facetdesc — Function

facetdesc(shapes::ShapeCollection)

Retrieve the shape type of the boundary facet.

MeshCore.nfacets — Function

nfacets(sd::AbstractShapeDesc{MD, NV, NF, FD}) where {MD, NV, NF, FD}

How many facets bound the shape?

nfacets(shapes::ShapeCollection)

Retrieve the number of boundary facets per shape.

MeshCore.facets — Function

facets(shapes::ShapeCollection)

Retrieve the connectivity of the facets.

MeshCore.facetconnectivity — Function

connectivity(shapes::ShapeCollection, i::I) where {I}

Retrieve connectivity of one shape from the collection.

Topology operations

MeshCore.skeleton — Function

skeleton(shapes::ShapeCollection; options...)

Compute the skeleton of the shape collection.

It consists of facets (shapes of manifold dimension one less than the manifold dimension of the shapes themselves).

Options

boundaryonly: include in the skeleton only shapes on the boundary of the input collection, true or false (default).

MeshCore.boundary — Function

boundary(shapes::ShapeCollection)

Compute the shape collection for the boundary of the collection on input.

This is a convenience version of the skeleton function.

MeshCore.increl — Function

increl(shapes::ShapeCollection, from::Int64, to::Int64)

Retrieve incidence relations of type from -> to.

MeshCore.increllist — Function

increllist(ir::IncRel, j::Int64)

Retrieve list of shapes incident on entity j.

Index

MeshCore.AbstractShapeDesc

MeshCore.H8ShapeDesc

MeshCore.IncRel

MeshCore.L2ShapeDesc

MeshCore.NoSuchShapeDesc

MeshCore.P1ShapeDesc

MeshCore.Q4ShapeDesc

MeshCore.ShapeCollection

MeshCore.ShapeCollection

MeshCore.T3ShapeDesc

MeshCore.T4ShapeDesc

MeshCore.Vertices

MeshCore.Vertices

MeshCore.boundary

MeshCore.connectivity

MeshCore.coordinates

MeshCore.coordinatetype

MeshCore.facetconnectivity

MeshCore.facetdesc

MeshCore.facets

MeshCore.increl

MeshCore.increllist

MeshCore.manifdim

MeshCore.nfacets

MeshCore.nshapes

MeshCore.nspacedims

MeshCore.nvertices

MeshCore.shapedesc

MeshCore.skeleton