Functions

Vertices

Missing docstring.

Missing docstring for Vertices(xyz::Array{T, 2}) where {T}. Check Documenter's build log for details.

MeshCore.nvertices — Function

nvertices(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many vertices does the shape connect?

nvertices(shapes::ShapeColl)

Retrieve the number of vertices per shape.

MeshCore.nspacedims — Function

nspacedims(loc::Locations{N, T}) where {N, T}

Retrieve the number of space dimensions.

MeshCore.coordinatetype — Function

coordinatetype(loc::Locations{N, T}) where {N, T}

Retrieve type of individual coordinates.

MeshCore.coordinates — Function

coordinates(Locations::Locations{N, T}, i::I) where {N, T, I<:Int}

Retrieve coordinates of a single vertex.

coordinates(loc::Locations{N, T}, I::SVector) where {N, T}

Access coordinates of selected Locations.

Shapes

Missing docstring.

Missing docstring for ShapeCollection(shapedesc::S, C::Array{T, 2}) where {S <: AbstractShapeDesc, T}. Check Documenter's build log for details.

MeshCore.shapedesc — Function

shapedesc(shapes::ShapeColl)

Retrieve the shape descriptor.

Missing docstring.

Missing docstring for connectivity. Check Documenter's build log for details.

MeshCore.nshapes — Function

nshapes(shapes::ShapeColl)

Number of shapes in the collection.

MeshCore.manifdim — Function

manifdim(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

What is the manifold dimension of this shape?

manifdim(shapes::ShapeColl)

Retrieve the manifold dimension of the collection.

Missing docstring.

Missing docstring for nvertices. Check Documenter's build log for details.

MeshCore.n1storderv — Function

n1storderv(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many 1st-order vertices are there per shape?

For instance, for hexahedra each shape has 8 1st-order vertices.

MeshCore.nshifts — Function

nshifts(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many circular shifts should be tried to figure out the orientation (sense)?

MeshCore.facetdesc — Function

facetdesc(shapes::ShapeColl)

Retrieve the shape type of the boundary facet.

MeshCore.nfacets — Function

nfacets(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many facets bound the shape?

nfacets(shapes::ShapeColl)

Retrieve the number of boundary facets per shape.

Missing docstring.

Missing docstring for facets. Check Documenter's build log for details.

MeshCore.edgetconnectivity — Function

edgetconnectivity(shapes::ShapeColl, j::Int64)

Retrieve connectivity of the j-th edget shape of the i-th shape from the collection.

MeshCore.edgetdesc — Function

edgetdesc(shapes::ShapeColl)

Retrieve the shape type of the boundary edget.

MeshCore.nedgets — Function

nedgets(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many edgets bound the shape?

nedgets(shapes::ShapeColl)

Retrieve the number of boundary edgets per shape.

Missing docstring.

Missing docstring for edgets. Check Documenter's build log for details.

Missing docstring.

Missing docstring for edgetconnectivity. Check Documenter's build log for details.

Incidence relations

MeshCore.nrelations — Function

nrelations(ir::IncRel)

Number of individual relations in the incidence relation.

MeshCore.nentities — Function

nentities(ir::IncRel, j::Int64)

Number of individual entities in the j-th relation in the incidence relation.

Missing docstring.

Missing docstring for (ir::IncRelFixed)(j::IT, k::IT) where {IT} = ir._v[j][k]. Check Documenter's build log for details.

Missing docstring.

Missing docstring for (ir::IncRelFixed)(j::IT) where {IT} = ir._v[j]. Check Documenter's build log for details.

Relations below the diagonal

MeshCore.skeleton — Function

skeleton(mesh::IncRel; options...)

Compute the skeleton of the mesh.

This computes a new incidence relation from an existing incidence relation.

Options

boundaryonly: include in the skeleton only shapes on the boundary of the input collection, true or false (default).

Returns

Incidence relation for the skeleton mesh. The left shape collection consists of facets (shapes of manifold dimension one less than the manifold dimension of the shapes themselves). The right shape collection is the same as for the input.

MeshCore.boundary — Function

boundary(mesh::IncRel)

Compute the incidence relation for the boundary of the collection on input.

This is a convenience version of the skeleton function.

MeshCore.boundedby — Function

boundedby(mesh::IncRel, fmesh::IncRel)

Compute the incidence relation d -> d-1 for d-dimensional shapes.

In other words, this is the incidence between shapes and the shapes that bound these shapes (facets). For tetrahedra as the shapes, the incidence relation lists the numbers of the faces that bound each individual tetrahedron. The resulting left shape is of the same shape description as in the mesh.

Note

The numbers of the facets are signed: positive when the facet bounds the shape in the sense in which it is defined by the shape as oriented with an outer normal; negative otherwise. The sense is defined by the numbering of the 1st-order vertices of the facet shape.

Returns

Incidence relation for the bounded-by mesh. The left shape collection is the same as for the mesh, the right shape collection is for the facets (shapes of manifold dimension one less than the manifold dimension of the shapes themselves).

MeshCore.boundedby2 — Function

boundedby2(mesh::IncRel, emesh::IncRel)

Compute the incidence relation d -> d-2 for d-dimensional shapes.

In other words, this is the incidence between shapes and the shapes that "bound" the boundaries of these shapes (i. e. edgets). For tetrahedra as the shapes, the incidence relation lists the numbers of the edges that "bound" each individual tetrahedron. The resulting shape is of the same shape description as the shapes on input.

Note

The numbers of the edgets are signed: positive when the edget bounds the shape in the sense in which it is defined by the shape as oriented with an outer normal; negative otherwise. The sense is defined by the numbering of the 1st-order vertices of the edget shape.

Returns

Incidence relation for the "bounded-by" mesh. The left shape collection is the same as for the mesh, the right shape collection is for the edgets (shapes of manifold dimension two less than the manifold dimension of the shapes themselves).

Relations above the diagonal

Missing docstring.

Missing docstring for increl_transpose. Check Documenter's build log for details.

Index

MeshCore.H8ShapeDesc

MeshCore.L2ShapeDesc

MeshCore.NoSuchShapeDesc

MeshCore.P1ShapeDesc

MeshCore.Q4ShapeDesc

MeshCore.T3ShapeDesc

MeshCore.T4ShapeDesc

MeshCore.boundary

MeshCore.boundedby

MeshCore.boundedby2

MeshCore.coordinates

MeshCore.coordinatetype

MeshCore.edgetconnectivity

MeshCore.edgetdesc

MeshCore.facetdesc

MeshCore.manifdim

MeshCore.n1storderv

MeshCore.nedgets

MeshCore.nentities

MeshCore.nfacets

MeshCore.nrelations

MeshCore.nshapes

MeshCore.nshifts

MeshCore.nspacedims

MeshCore.nvertices

MeshCore.shapedesc

MeshCore.skeleton