Functions

Shapes

MeshCore.ShapeColl — Method

ShapeColl(shapedesc::S, nshapes::Int64) where {S <: AbsShapeDesc}

Set up new shape collection with an empty dictionary of attributes and a default name.

MeshCore.ShapeColl — Method

ShapeColl(shapedesc::S, nshapes::Int64, d::Dict{Symbol, Any}) where {S <: AbsShapeDesc}

Set up new shape collection with a default name.

MeshCore.ShapeColl — Method

ShapeColl(shapedesc::S, nshapes::Int64, d::Dict{Symbol, Any}) where {S <: AbsShapeDesc}

Set up new shape collection with an empty dictionary of attributes.

MeshCore.shapedesc — Function

shapedesc(shapes::ShapeColl)

Retrieve the shape descriptor.

MeshCore.nshapes — Function

nshapes(shapes::ShapeColl)

Number of shapes in the collection.

MeshCore.manifdim — Function

manifdim(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

What is the manifold dimension of this shape?

manifdim(shapes::ShapeColl)

Retrieve the manifold dimension of the collection.

MeshCore.nvertices — Function

nvertices(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many vertices does the shape connect?

nvertices(shapes::ShapeColl)

Retrieve the number of vertices per shape.

MeshCore.facetdesc — Function

facetdesc(shapes::ShapeColl)

Retrieve the shape type of the boundary facet.

MeshCore.nfacets — Function

nfacets(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many facets bound the shape?

nfacets(shapes::ShapeColl)

Retrieve the number of boundary facets per shape.

MeshCore.facetconnectivity — Function

facetconnectivity(shapes::ShapeColl, j::Int64)

Retrieve connectivity of the j-th facet shape of the i-th shape from the collection.

MeshCore.ridgedesc — Function

ridgedesc(shapes::ShapeColl)

Retrieve the shape type of the boundary ridge.

MeshCore.nridges — Function

nridges(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many ridges bound the shape?

nridges(shapes::ShapeColl)

Retrieve the number of boundary ridges per shape.

MeshCore.ridgeconnectivity — Function

ridgeconnectivity(shapes::ShapeColl, j::Int64)

Retrieve connectivity of the j-th ridge shape of the i-th shape from the collection.

MeshCore.n1storderv — Function

n1storderv(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many 1st-order vertices are there per shape?

For instance, for hexahedra each shape has 8 1st-order vertices.

MeshCore.nshifts — Function

nshifts(sd::AbsShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many circular shifts should be tried to figure out the orientation (sense)?

MeshCore.attribute — Function

attribute(shapes::ShapeColl, s::Symbol)

Retrieve a named attribute.

Incidence relations

MeshCore.IncRel — Method

IncRel(left::ShapeColl{LEFT}, right::ShapeColl{RIGHT}, data::Matrix{MT}) where {LEFT<:AbsShapeDesc, RIGHT<:AbsShapeDesc, MT}

Convenience constructor with a vector of vectors and a default name.

MeshCore.IncRel — Method

IncRel(left::ShapeColl{LEFT}, right::ShapeColl{RIGHT}, data::Matrix{MT}) where {LEFT<:AbsShapeDesc, RIGHT<:AbsShapeDesc, MT}

Convenience constructor supplying a matrix instead of a vector of vectors and a default name.

MeshCore.indextype — Function

indextype(ir::IncRel)

Provide type of the incidence index (entity number). Presumably some type of integer.

MeshCore.nrelations — Function

nrelations(ir::IncRel)

Number of individual relations in the incidence relation.

MeshCore.nentities — Function

nentities(ir::IncRel, j::IT) where {IT}

Number of individual entities in the j-th relation in the incidence relation.

MeshCore.retrieve — Function

retrieve(ir::IncRel{LEFT, RIGHT, T}, j::IT1, k::IT2) where {LEFT<:AbsShapeDesc, RIGHT<:AbsShapeDesc, T, IT1, IT2}

Retrieve the incidence relation for j-th relation, k-th entity.

retrieve(ir::IncRel{LEFT, RIGHT, T}, j::IT) where {LEFT<:AbsShapeDesc, RIGHT<:AbsShapeDesc, T, IT}

Retrieve the row of the incidence relation for j-th relation.

Relations below the diagonal

MeshCore.skeleton — Function

skeleton(ir::IncRel, name = "skt")

Compute the skeleton of an incidence relation.

This computes a new incidence relation from an existing incidence relation.

Options

boundaryonly: include in the skeleton only shapes on the boundary of the input collection, true or false (default).

Returns

Incidence relation for the skeleton of the input incidence relation. The left shape collection consists of facets (shapes of manifold dimension one less than the manifold dimension of the shapes themselves). The right shape collection is the same as for the input.

MeshCore.boundary — Function

boundary(ir::IncRel, name = "bdr")

Compute the incidence relation for the boundary of the incidence relation on input.

The skeleton function.

MeshCore.bbyfacets — Function

bbyfacets(ir::IncRel, fir::IncRel, tfir::IncRel, name = "bbf")

Compute the incidence relation (d, d-1) for d-dimensional shapes.

In other words, this is the incidence relation between shapes and the shapes that bound these shapes (facets). For tetrahedra as the shapes, the incidence relation lists the numbers of the faces that bound each individual tetrahedron. The resulting left shape is of the same shape description as in the ir.

Arguments

ir = incidence relation (d, 0),

fir = incidence relation (d-1, 0),

tfir = transpose of the above, incidence relation (0, d-1).

Returns

Incidence relation for the bounded-by relationship between shape collections. The left shape collection is the same as for the ir, the right shape collection is for the facets (shapes of manifold dimension one less than the manifold dimension of the shapes themselves).

Note

The numbers of the facets are signed: positive when the facet bounds the shape in the sense in which it is defined by the shape as oriented with an outer normal; negative otherwise. The sense is defined by the numbering of the 1st-order vertices of the facet shape.

bbyfacets(ir::IncRel, fir::IncRel, name = "bbf")

Compute the incidence relation (d, d-1) for d-dimensional shapes.

Convenience function where the transpose of the incidence relation on the right is computed on the fly.

See also: bbyfacets(ir::IncRel, fir::IncRel, tfir::IncRel)

MeshCore.bbyridges — Function

bbyridges(ir::IncRel, eir::IncRel, teir::IncRel, name = "bbr")

Compute the incidence relation (d, d-2) for d-dimensional shapes.

In other words, this is the incidence between shapes and the shapes that "bound" the boundaries of these shapes (i. e. ridges). For tetrahedra as the shapes, the incidence relation lists the numbers of the edges that "bound" each individual tetrahedron. The resulting shape is of the same shape description as the shapes on input.

Returns

The incidence relation for the "bounded by" relationship between shape collections. The left shape collection is the same as for the ir, the right shape collection is for the ridges (shapes of manifold dimension two less than the manifold dimension of the shapes themselves).

Note

The numbers of the ridges are signed: positive when the ridge bounds the shape in the sense in which it is defined by the shape as oriented with an outer normal; negative otherwise. The sense is defined by the numbering of the 1st-order vertices of the ridge shape.

bbyridges(ir::IncRel, eir::IncRel, name = "bbr")

Compute the incidence relation (d, d-2) for d-dimensional shapes.

Convenience function where the transpose of the incidence relation on the right is computed on the fly.

See also: bbyridges(ir::IncRel, eir::IncRel, efir::IncRel)

Relations above the diagonal

MeshCore.transpose — Method

transpose(ir::IncRel, name = "trp")

Compute the incidence relation (d1, d2), where d2 >= d1.

This only makes sense for d2 >= 0. For d2=1 we get for each vertex the list of edges connected to the vertex, and analogously faces and cells for d=2 and d=3.

Returns

Incidence relation for the transposed incidence relation. The left and right shape collection are swapped in the output relative to the input.

Attributes

MeshCore.nvals — Method

nvals(d::AttribDataWrapper{T})

How many values are there stored in this wrapper?

MeshCore.val — Method

val(d::AttribDataWrapper{T}, j::Int64) where {T}

Access the j-th value of the wrapper.

Missing docstring.

Missing docstring for (d::AttribDataWrapper{T})(j::Int64) where {T}. Check Documenter's build log for details.

Index

MeshCore.SHAPE_DESC

MeshCore.AbsAttrib

MeshCore.AbsShapeDesc

MeshCore.Attrib

MeshCore.AttribDataWrapper

MeshCore.H8ShapeDesc

MeshCore.IncRel

MeshCore.IncRel

MeshCore.IncRel

MeshCore.L2ShapeDesc

MeshCore.NoSuchShapeDesc

MeshCore.P1ShapeDesc

MeshCore.Q4ShapeDesc

MeshCore.ShapeColl

MeshCore.ShapeColl

MeshCore.ShapeColl

MeshCore.ShapeColl

MeshCore.T3ShapeDesc

MeshCore.T4ShapeDesc

MeshCore.attribute

MeshCore.bbyfacets

MeshCore.bbyridges

MeshCore.boundary

MeshCore.facetconnectivity

MeshCore.facetdesc

MeshCore.indextype

MeshCore.manifdim

MeshCore.n1storderv

MeshCore.nentities

MeshCore.nfacets

MeshCore.nrelations

MeshCore.nridges

MeshCore.nshapes

MeshCore.nshifts

MeshCore.nvals

MeshCore.nvertices

MeshCore.retrieve

MeshCore.ridgeconnectivity

MeshCore.ridgedesc

MeshCore.shapedesc

MeshCore.skeleton

MeshCore.transpose

MeshCore.val