Functions

Vertices

MeshCore.Vertices — Method

Vertices(xyz::Array{T, 2}) where {T}

Construct from a matrix (2D array), one vertex per row.

MeshCore.nvertices — Function

nvertices(v::Vertices{N, T}) where {N, T}

Retrieve number of vertices.

nvertices(sd::AbstractShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many vertices does the shape connect?

nvertices(shapes::ShapeCollection)

Retrieve the number of vertices per shape.

MeshCore.nspacedims — Function

nspacedims(v::Vertices{N, T}) where {N, T}

Retrieve the number of space dimensions.

MeshCore.coordinatetype — Function

coordinatetype(v::Vertices{N, T}) where {N, T}

Retrieve type of individual coordinates.

MeshCore.coordinates — Function

coordinates(vertices::Vertices{N, T}, i::I) where {N, T, I<:Int}

Retrieve coordinates of a single vertex.

coordinates(vertices::Vertices{N, T}, I::SVector) where {N, T}

Access coordinates of selected vertices.

Shapes

MeshCore.ShapeCollection — Method

ShapeCollection(shapedesc::S, C::Array{T, 2}) where {S <: AbstractShapeDesc, T}

Convenience constructor from a matrix. One shape per row.

MeshCore.shapedesc — Function

shapedesc(shapes::ShapeCollection)

Retrieve the shape descriptor.

MeshCore.connectivity — Function

connectivity(shapes::ShapeCollection, i::IT) where {IT}

Retrieve connectivity of the i-th shape from the collection.

connectivity(shapes::ShapeCollection, I::SVector)

Retrieve connectivity of multiple shapes from the collection.

Static arrays are used to help the compiler avoid memory allocation.

MeshCore.nshapes — Function

nshapes(shapes::ShapeCollection)

Number of shapes in the collection.

MeshCore.manifdim — Function

manifdim(sd::AbstractShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

What is the manifold dimension of this shape?

manifdim(shapes::ShapeCollection)

Retrieve the manifold dimension of the collection.

Missing docstring.

Missing docstring for nvertices. Check Documenter's build log for details.

MeshCore.n1storderv — Function

n1storderv(sd::AbstractShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many 1st-order vertices are there per shape?

For instance, for hexahedra each shape has 8 1st-order vertices.

MeshCore.nshifts — Function

nshifts(sd::AbstractShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many circular shifts should be tried to figure out the orientation (sense)?

MeshCore.facetdesc — Function

facetdesc(shapes::ShapeCollection)

Retrieve the shape type of the boundary facet.

MeshCore.nfacets — Function

nfacets(sd::AbstractShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many facets bound the shape?

nfacets(shapes::ShapeCollection)

Retrieve the number of boundary facets per shape.

Missing docstring.

Missing docstring for facets. Check Documenter's build log for details.

MeshCore.edgetconnectivity — Function

edgetconnectivity(shapes::ShapeCollection, i::I, j::I) where {I}

Retrieve connectivity of the j-th edget shape of the i-th shape from the collection.

MeshCore.edgetdesc — Function

edgetdesc(shapes::ShapeCollection)

Retrieve the shape type of the boundary edget.

MeshCore.nedgets — Function

nedgets(sd::AbstractShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}) where {MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

How many edgets bound the shape?

nedgets(shapes::ShapeCollection)

Retrieve the number of boundary edgets per shape.

Missing docstring.

Missing docstring for edgets. Check Documenter's build log for details.

Missing docstring.

Missing docstring for edgetconnectivity. Check Documenter's build log for details.

Incidence relations

Missing docstring.

Missing docstring for nrelations. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nentities. Check Documenter's build log for details.

Missing docstring.

Missing docstring for (ir::IncRelFixed)(j::IT, k::IT) where {IT} = ir._v[j][k]. Check Documenter's build log for details.

Missing docstring.

Missing docstring for (ir::IncRelFixed)(j::IT) where {IT} = ir._v[j]. Check Documenter's build log for details.

Relations below the diagonal

MeshCore.skeleton — Function

skeleton(shapes::ShapeCollection; options...)

Compute the skeleton of the shape collection.

This computes a new shape collection from an existing shape collection. It consists of facets (shapes of manifold dimension one less than the manifold dimension of the shapes themselves).

Options

boundaryonly: include in the skeleton only shapes on the boundary of the input collection, true or false (default).

MeshCore.boundary — Function

boundary(shapes::ShapeCollection)

Compute the shape collection for the boundary of the collection on input.

This is a convenience version of the skeleton function.

MeshCore.boundedby — Function

boundedby(shapes::ShapeCollection, facetshapes::ShapeCollection)

Compute the shape collection that expresses the incidence d -> d-1 for d-dimensional shapes.

In other words, this is the incidence between shapes and the shapes that bound these shapes (facets). For tetrahedra as the shapes, the incidence relation lists the numbers of the faces that bound each individual tetrahedron. The resulting shape is of the same shape description as the shapes on input.

Note

The numbers of the facets are signed: positive when the facet bounds the shape in the sense in which it is defined by the shape as oriented with an outer normal; negative otherwise. The sense is defined by the numbering of the 1st-order vertices of the facet shape.

MeshCore.boundedby2 — Function

boundedby2(shapes::ShapeCollection, edgetshapes::ShapeCollection)

Compute the shape collection that expresses the incidence d -> d-2 for d-dimensional shapes.

In other words, this is the incidence between shapes and the shapes that "bound" the boundaries of these shapes (i. e. edgets). For tetrahedra as the shapes, the incidence relation lists the numbers of the edges that "bound" each individual tetrahedron. The resulting shape is of the same shape description as the shapes on input.

Note

The numbers of the edgets are signed: positive when the edget bounds the shape in the sense in which it is defined by the shape as oriented with an outer normal; negative otherwise. The sense is defined by the numbering of the 1st-order vertices of the edget shape.

Relations above the diagonal

MeshCore.increl_transpose — Function

increl_transpose(ir::IR) where {IR}

Compute the incidence relation d1 -> d2, where d2 >= d1 for d2-dimensional shapes.

This only makes sense for d2 >= 0. For d2=1 we get for each vertex the list of edges connected to the vertex, and analogously faces and cells for d=2 and d=3.

Index

MeshCore.AbstractShapeDesc

MeshCore.H8ShapeDesc

MeshCore.IncRelFixed

MeshCore.IncRelVar

MeshCore.L2ShapeDesc

MeshCore.NoSuchShapeDesc

MeshCore.P1ShapeDesc

MeshCore.Q4ShapeDesc

MeshCore.ShapeCollection

MeshCore.ShapeCollection

MeshCore.T3ShapeDesc

MeshCore.T4ShapeDesc

MeshCore.Vertices

MeshCore.Vertices

MeshCore.boundary

MeshCore.boundedby

MeshCore.boundedby2

MeshCore.connectivity

MeshCore.coordinates

MeshCore.coordinatetype

MeshCore.edgetconnectivity

MeshCore.edgetdesc

MeshCore.facetdesc

MeshCore.increl_transpose

MeshCore.manifdim

MeshCore.n1storderv

MeshCore.nedgets

MeshCore.nfacets

MeshCore.nshapes

MeshCore.nshifts

MeshCore.nspacedims

MeshCore.nvertices

MeshCore.shapedesc

MeshCore.skeleton