Types

Vertices

MeshCore.Vertices — Type

Vertices{N, T}

Vertices of the mesh.

N = number of coordinates, equal to the number of space dimensions,

T = type of the individual coordinates (concrete subtype of AbstractFloat).

Shapes

Shape descriptors

MeshCore.AbstractShapeDesc — Type

AbstractShapeDesc

Abstract shape descriptor (simplex, or cuboid, of various manifold dimension)

The definitions of concrete types below define a set of shape descriptors, covering the usual element shapes: point, line segment, quadrilateral, hexahedron, and a similar hierarchy for the simplex elements: point, line segment, triangle, tetrahedron.

The concrete types of the shape descriptor will provide access to manifdim = manifold dimension of the shape (0, 1, 2, or 3), nnodes = number of nodes (vertices) of the shape, facetdesc = shape descriptor of the shapes on the boundary of this shape, nfacets = number of shapes on the boundary of this shape, facets = definitions of the shapes on the boundary of this shape in terms of local connectivity

MeshCore.NoSuchShapeDesc — Type

NoSuchShapeDesc

Base description: no shape is of this description.

MeshCore.P1ShapeDesc — Type

P1ShapeDesc{BS}

Shape description of a point shape. BS is the type of the facet (boundary shape).

MeshCore.L2ShapeDesc — Type

L2ShapeDesc{BS}

Shape description of a line segment shape. BS is the type of the facet (boundary shape).

MeshCore.Q4ShapeDesc — Type

Q4ShapeDesc{BS}

Shape description of a quadrilateral shape. BS is the type of the facet (boundary shape).

MeshCore.H8ShapeDesc — Type

H8ShapeDesc{BS}

Shape description of a hexahedral shape. BS is the type of the facet (boundary shape).

MeshCore.T3ShapeDesc — Type

T3ShapeDesc{BS}

Shape description of a triangular shape. BS is the type of the facet (boundary shape).

MeshCore.T4ShapeDesc — Type

T4ShapeDesc{BS}

Shape description of a tetrahedral shape. BS is the type of the facet (boundary shape).

Collection of shapes

MeshCore.ShapeCollection — Type

ShapeCollection{S <: AbstractShapeDesc, N, T}

This is the type of a homogeneous collection of finite element shapes.

S = shape descriptor:

N = number of vertices incident on the shape (i. e. by which the shape is defined)

T = type of the vertex numbers (some concrete type of Integer).

Incidence relations

MeshCore.VertexToShape — Type

VertexToShape(vertices, shapes::ShapeCollection{S, N, T}) where {S, N, T}

Compute the incident relation 0 -> d for d-dimensional shapes.

This only makes sense for d > 0. For d=1 we get for each vertex the list of edges connected to the vertex, and analogously faces and cells for d=2 and d=3.