Types

Shape descriptors

MeshCore.AbsShapeDesc — Type

AbsShapeDesc

Abstract shape descriptor (simplex, or cuboid, of various manifold dimension)

The definitions of concrete types below define a set of shape descriptors, covering the usual element shapes: point, line segment, quadrilateral, hexahedron, and a similar hierarchy for the simplex elements: point, line segment, triangle, tetrahedron.

The concrete types of the shape descriptor will provide access to the following properties of the shape S:

manifdim = manifold dimension of the shape S (i. e. 0, 1, 2, or 3),

nvertices = number of vertices of the shape S,

n1storderv = number of first-order vertices for the shape S, for instance a 6-node triangle has 3 first-order vertices,

nfacets = number of shapes on the boundary of the shape S,

facetdesc = shape descriptor of the shapes on the boundary of the shape S,

facets = definitions of the shapes on the boundary of the shape S in terms of local connectivity

ridgedesc = shape descriptor of the shapes on the boundary of the boundary of the shape S (which we call ridges here),

ridges = definitions of the shapes on the boundary of the boundary of the shape S in terms of local connectivity

The bit-type values are defined with the type parameters:

MD is the manifold dimension,

NV is the number of connected vertices,

NF is the number of boundary facets,

FD is the descriptor of the facet (boundary shape).

NR is the number of boundary ridges,

RD is the descriptor of the ridge (boundary shape).

NFOV is number of first-order vertices,

NSHIFTS is the number of shifts that should be attempted when matching shapes to determine orientation.

MeshCore.NoSuchShapeDesc — Type

NoSuchShapeDesc

Base descriptor: no shape is of this description.

MeshCore.P1ShapeDesc — Type

P1ShapeDesc{MD, NV, NF, FD, NR, RD, NFOV, NSHIFTS}

Shape descriptor of a point shape.

MeshCore.L2ShapeDesc — Type

L2ShapeDesc{MD, NV, NF, FD, NR, RD, NFOV, NSHIFTS}

Shape descriptor of a line segment shape.

MeshCore.Q4ShapeDesc — Type

Q4ShapeDesc{MD, NV, NF, FD, NR, RD, NFOV, NSHIFTS}

Shape descriptor of a quadrilateral shape.

Facets: First the two orthogonal to xi, then the two orthogonal to eta. The coordinate along the edge increases.

Ridges: Counterclockwise.

MeshCore.H8ShapeDesc — Type

H8ShapeDesc{MD, NV, NF, FD, NR, RD, NFOV, NSHIFTS}

Shape descriptor of a hexahedral shape.

Facets: First the two faces orthogonal to xi, then the two orthogonal to eta, finally the two orthogonal to zeta. The vertices are given counterclockwise when looking from the outside.

Ridges: First the four edges parallel to xi, then the four parallel to eta, and finally the four parallel to zeta.

MeshCore.T3ShapeDesc — Type

T3ShapeDesc{MD, NV, NF, FD, NR, RD, NFOV, NSHIFTS}

Shape descriptor of a triangular shape.

Facets: We start with the edge opposite to vertex 1, then the edge opposite to vertex 2 and so on. The vertices define counterclockwise orientation.

Ridges: The vertices in counterclockwise order.

MeshCore.T4ShapeDesc — Type

T4ShapeDesc{MD, NV, NF, FD, NR, RD, NFOV, NSHIFTS}

Shape descriptor of a tetrahedral shape.

Facets: We start with the face opposite to vertex 1, then the face opposite to vertex 2 and so on. The vertices define counterclockwise orientation when viewed from outside of the shape.

Ridges: The vertices of the edges are defined in the order based on the right-hand side rotation rule: rotate about an edge (ridge) as if it was a rotation vector with positive orientation. The vertices on the edge opposite across the tetrahedron are traversed in the order given by the rotation.

MeshCore.SHAPE_DESC — Constant

SHAPE_DESC

Dictionary of all the descriptors.

Shape collections

MeshCore.ShapeColl — Type

ShapeColl{S <: AbsShapeDesc}

This is the type of a homogeneous collection of finite element shapes.

S = shape descriptor: subtype of AbsShapeDesc.

Attributes

MeshCore.AbsAttrib — Type

AbsAttrib

Abstract type of attribute. It is a subtype of AbstractArray.

Missing docstring.

Missing docstring for Attrib{F}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AttribDataWrapper{T}. Check Documenter's build log for details.

Incidence relations

MeshCore.IncRel — Type

IncRel{LEFT<:AbsShapeDesc, RIGHT<:AbsShapeDesc, T}

Incidence relation expressing connectivity between an entity of the shape on the left and a list of entities of the shape on the right.

The incidence relation may be referred to by its code: (d1, d2), where d1 is the manifold dimension of the shape collection on the left, and d2 is the manifold dimension of the shape collection on the right.