Types

Geometry

MeshCore.Locations — Type

Locations{N, T}

Locations of the vertices.

N = number of coordinates, equal to the number of space dimensions,

T = type of the individual coordinates (concrete subtype of AbstractFloat).

Shape descriptors

MeshCore.AbsShapeDesc — Type

AbsShapeDesc

Abstract shape descriptor (simplex, or cuboid, of various manifold dimension)

The definitions of concrete types below define a set of shape descriptors, covering the usual element shapes: point, line segment, quadrilateral, hexahedron, and a similar hierarchy for the simplex elements: point, line segment, triangle, tetrahedron.

The concrete types of the shape descriptor will provide access to manifdim = manifold dimension of the shape (0, 1, 2, or 3), nvertices = number of vertices of the shape, n1storderv = number of first-order vertices for the shape, for instance a 6-node triangle has 3 first-order vertices, nfacets = number of shapes on the boundary of this shape, facetdesc = shape descriptor of the shapes on the boundary of this shape, facets = definitions of the shapes on the boundary of this shape in terms of local connectivity

The bit-type values are defined with the type parameters: MD is the manifold dimension, NV is the number of connected vertices, NF is the number of boundary facets, N1OV number of first-order vertices, FD is the descriptor of the facet (boundary shape).

MeshCore.NoSuchShapeDesc — Type

NoSuchShapeDesc

Base descriptor: no shape is of this description.

MeshCore.P1ShapeDesc — Type

P1ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a point shape.

MeshCore.L2ShapeDesc — Type

L2ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a line segment shape.

MeshCore.Q4ShapeDesc — Type

Q4ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a quadrilateral shape.

MeshCore.H8ShapeDesc — Type

H8ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a hexahedral shape.

MeshCore.T3ShapeDesc — Type

T3ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a triangular shape.

MeshCore.T4ShapeDesc — Type

T4ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a tetrahedral shape.

Shape collections

MeshCore.ShapeColl — Type

ShapeColl{S <: AbsShapeDesc}

This is the type of a homogeneous collection of finite element shapes.

S = shape descriptor: subtype of AbsShapeDesc.

Attributes

MeshCore.AbsAttrib — Type

AbsAttrib

Abstract type of attribute.

Missing docstring.

Missing docstring for Attrib{F}<:AbsAttrib. Check Documenter's build log for details.

Incidence relations

MeshCore.IncRel — Type

IncRel{LEFT<:AbsShapeDesc, RIGHT<:AbsShapeDesc, T}

Incidence relation expressing connectivity between an entity of the shape on the left and a list of entities of the shape on the right.