Types

Geometry

MeshCore.Locations — Type

Locations{N, T}

Locations of the vertices.

N = number of coordinates, equal to the number of space dimensions,

T = type of the individual coordinates (concrete subtype of AbstractFloat).

MeshCore.LocAccess — Type

LocAccess{N, T}

Data structure to access the locations of vertices.

The access to the coordinates may be either through a function coordinates, or through a "callable object" interface as

(la::LocAccess{N, T})(j::Int64) where {N, T}

Refer to the example below.

Example

The use is intended for the data stored in an attribute:

using MeshCore: Locations, LocAccess, Attrib, coordinates
xyz = [30.0 3.0; 633.3 800.0]
locs =  Locations(xyz)
la = LocAccess(locs)
a = Attrib(la)
using Test
@test a.val(2) == [633.3, 800.0]
@test coordinates(a.val, 2) == [633.3, 800.0]

Shape descriptors

MeshCore.AbsShapeDesc — Type

AbsShapeDesc

Abstract shape descriptor (simplex, or cuboid, of various manifold dimension)

The definitions of concrete types below define a set of shape descriptors, covering the usual element shapes: point, line segment, quadrilateral, hexahedron, and a similar hierarchy for the simplex elements: point, line segment, triangle, tetrahedron.

The concrete types of the shape descriptor will provide access to manifdim = manifold dimension of the shape (0, 1, 2, or 3), nvertices = number of vertices of the shape, n1storderv = number of first-order vertices for the shape, for instance a 6-node triangle has 3 first-order vertices, nfacets = number of shapes on the boundary of this shape, facetdesc = shape descriptor of the shapes on the boundary of this shape, facets = definitions of the shapes on the boundary of this shape in terms of local connectivity

The bit-type values are defined with the type parameters: MD is the manifold dimension, NV is the number of connected vertices, NF is the number of boundary facets, N1OV number of first-order vertices, FD is the descriptor of the facet (boundary shape).

MeshCore.NoSuchShapeDesc — Type

NoSuchShapeDesc

Base descriptor: no shape is of this description.

MeshCore.P1ShapeDesc — Type

P1ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a point shape.

MeshCore.L2ShapeDesc — Type

L2ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a line segment shape.

MeshCore.Q4ShapeDesc — Type

Q4ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a quadrilateral shape.

MeshCore.H8ShapeDesc — Type

H8ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a hexahedral shape.

MeshCore.T3ShapeDesc — Type

T3ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a triangular shape.

MeshCore.T4ShapeDesc — Type

T4ShapeDesc{MD, NV, NF, FD, NE, ED, N1OV, NSHIFTS}

Shape descriptor of a tetrahedral shape.

MeshCore.SHAPE_DESC — Constant

SHAPE_DESC

Dictionary of all the descriptors.

Shape collections

MeshCore.ShapeColl — Type

ShapeColl{S <: AbsShapeDesc}

This is the type of a homogeneous collection of finite element shapes.

S = shape descriptor: subtype of AbsShapeDesc.

Attributes

MeshCore.AbsAttrib — Type

AbsAttrib

Abstract type of attribute.

MeshCore.Attrib — Type

Attrib{F}<:AbsAttrib

Attribute: F is either a function type, or type of a callable object. The value is linked to the serial number of the object.

Example

Attribute to allow access to the locations of the vertices.

la = LocAccess(locs)
a = Attrib(la)
vertices = ShapeColl(P1, size(xyz, 1), Dict(:geom=>a))
a = vertices.attributes[:geom]
@test a.val(10) == [633.3333333333334, 800.0]

Attribute to label the vertices.

a = Attrib(i -> 1)
vertices = ShapeColl(P1, size(xyz, 1), Dict(:label=>a))
a = vertices.attributes[:label]
@test a.val(10) == 1

Incidence relations

MeshCore.IncRel — Type

IncRel{LEFT<:AbsShapeDesc, RIGHT<:AbsShapeDesc, T}

Incidence relation expressing connectivity between an entity of the shape on the left and a list of entities of the shape on the right.